adding some security measures

+++ app/Http/Controllers/PostController.php
@@ -154,6 +154,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id != auth()->id()) {
+            abort(403);
+        }
         $post->delete();
         return redirect('/@'.auth()->user()->username);
     }
